refactor : Join 애그리거트명을 Belong으로 변경한다
