+++ backend/src/main/kotlin/com/dclass/backend/application/PostValidator.kt
@@ -1,7 +1,7 @@
 package com.dclass.backend.application
 
 import com.dclass.backend.domain.community.CommunityRepository
-import com.dclass.backend.domain.join.JoinRepository
+import com.dclass.backend.domain.belong.BelongRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -9,12 +9,12 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 @Transactional
 class PostValidator(
-    private val joinRepository: JoinRepository,
+    private val belongRepository: BelongRepository,
     private val communityRepository: CommunityRepository
 ) {
     fun validateCreatePost(userId: Long, communityId: Long) {
-        val join = joinRepository.findByUserId(userId)
+        val belong = belongRepository.findByUserId(userId)
         val community = communityRepository.findByIdOrNull(communityId)
-        check(join.contain(communityId)) { "해당 커뮤니티에 게시글을 작성할 수 없습니다." }
+        check(belong.contain(communityId)) { "해당 커뮤니티에 게시글을 작성할 수 없습니다." }
     }
 }
