+++ backend/src/main/kotlin/com/dclass/backend/domain/belong/BelongRepository.kt
@@ -1,0 +1,7 @@
+package com.dclass.backend.domain.belong
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface BelongRepository : JpaRepository<Belong, Long> {
+    fun findByUserId(userId: Long): Belong
+}
