--- backend/src/main/kotlin/com/dclass/backend/domain/join/JoinRepository.kt
@@ -1,7 +1,0 @@
-package com.dclass.backend.domain.join
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface JoinRepository : JpaRepository<Join, Long> {
-    fun findByUserId(userId: Long): Join
-}
